+++ database/migrations/2021_10_16_181613_create_like_lessons.php
@@ -15,11 +15,11 @@ class CreateLikeLessons extends Migration
     {
         Schema::create('like_lessons', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_user')->unsigned();
+            $table->bigInteger('id_user')->unsigned();
             $table->foreign('id_user')->references('id')->on('users')->cascadeOnDelete();
-            $table->integer('id_lesson')->unsigned()->nullable();
+            $table->bigInteger('id_lesson')->unsigned()->nullable();
             $table->foreign('id_lesson')->references('id')->on('lessons')->cascadeOnDelete();
-            $table->integer('id_cmt')->unsigned()->nullable();
+            $table->bigInteger('id_cmt')->unsigned()->nullable();
             $table->foreign('id_cmt')->references('id')->on('cmt_lessons')->cascadeOnDelete();
             $table->timestamps();
         });
