+++ database/migrations/2021_10_16_183736_create_like_topics.php
@@ -15,11 +15,11 @@ class CreateLikeTopics extends Migration
     {
         Schema::create('like_topics', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_user')->unsigned();
+            $table->bigInteger('id_user')->unsigned();
             $table->foreign('id_user')->references('id')->on('users')->cascadeOnDelete();
-            $table->integer('id_topic')->unsigned()->nullable();
+            $table->bigInteger('id_topic')->unsigned()->nullable();
             $table->foreign('id_topic')->references('id')->on('topics')->cascadeOnDelete();
-            $table->integer('id_cmt')->unsigned()->nullable();
+            $table->bigInteger('id_cmt')->unsigned()->nullable();
             $table->foreign('id_cmt')->references('id')->on('cmt_topics')->cascadeOnDelete();
             $table->timestamps();
         });
